requirements.txt
discord.py
json
requests
BeautifulSoup4
gspread
oauth2client
PyNaCl
